mnist_models/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
scikit-learn
matplotlib
flask
numpy
pandas
Pillow